[TCR] feat: Add functionality to retrieve messages mentioning a user
Introduces a new function `getMentionsForUser` in `TimelineService`. This function iterates through all messages stored in the service, identifies messages that contain a mention (preceded by '@') of the specified username, and returns a list of these messages sorted by timestamp in descending order.

A new test case `should find messages where a user is mentioned` has been added to `TimelineServiceTest` to verify the correct behavior of the new function, including finding mentions for different users and ensuring the results are sorted chronologically.